ETL/requirements.txt
pyspark==3.5.6
pandas
pyarrow
datetime
python-dotenv
pymysql
